+++ lib/fog/dnsimple/dns.rb
@@ -2,8 +2,8 @@ require "fog/core"
 require "fog/json"
 
 module Fog
-  module DNS
-    class Dnsimple < Fog::Service
+  module Dnsimple
+    class DNS < Fog::Service
       recognizes :dnsimple_token, :dnsimple_account, :dnsimple_url
 
       model_path 'fog/dnsimple/models/dns'
@@ -130,5 +130,12 @@ module Fog
         end
       end
     end
+  end
+end
+
+module Fog
+  module DNS
+    # Keeps the constant from the service::provider format for compatibility.
+    Dnsimple = Fog::Dnsimple::DNS
   end
 end