+++ lib/fog/dnsimple/requests/dns/update_record.rb
@@ -1,6 +1,6 @@
 module Fog
-  module DNS
-    class Dnsimple
+  module Dnsimple
+    class DNS
       class Real
         # Update the given record for the given domain.
         #